modularise linalg ops (#3464)
DEV
